data/stream/requirements.txt
fastapi[standard]
kafka-python
python-dotenv